// src/enums/status.enum.ts
export enum DoctorStatusEnum {
  Invited = "invited",
  Accepted = "accepted",
  Active = "active",
  Pending = "pending",
  Declined = "declined",
  Inactive = "inactive",
}

export enum ClientStatusEnum {
  Deactivated = 0,
  Available = 1,
}

export enum ClinicalSpaPaymentEnum {
  Due = "due",
  Pending = "pending",
  Paid = "paid",
}